+++ compiler-tests/src/test/kotlin/dev/zacsweers/metro/compiler/MetroDirectives.kt
@@ -29,6 +29,7 @@ object MetroDirectives : SimpleDirectivesContainer() {
 
   // Dependency directives.
   val WITH_ANVIL by directive("Add Anvil as dependency and configure custom annotations.")
+  val WITH_KI_ANVIL by directive("Add kotlin-inject-nnvil as dependency and configure custom annotations.")
   val WITH_DAGGER by directive("Add Dagger as dependency and configure custom annotations.")
   val ENABLE_DAGGER_INTEROP by
     directive("Enable Dagger interop. This implicitly applies WITH_DAGGER directive as well.")