+++ compiler-tests/src/test/kotlin/dev/zacsweers/metro/compiler/interop/AnvilRuntimeTestServices.kt
@@ -16,6 +16,11 @@ private val anvilRuntimeClasspath =
   System.getProperty("anvilRuntime.classpath")?.split(File.pathSeparator)?.map(::File)
     ?: error("Unable to get a valid classpath from 'anvilRuntime.classpath' property")
 
+
+private val kiAnvilRuntimeClasspath =
+  System.getProperty("kiAnvilRuntime.classpath")?.split(File.pathSeparator)?.map(::File)
+    ?: error("Unable to get a valid classpath from 'kiAnvilRuntime.classpath' property")
+
 fun TestConfigurationBuilder.configureAnvilAnnotations() {
   useConfigurators(::AnvilRuntimeEnvironmentConfigurator)
   useCustomRuntimeClasspathProviders(::AnvilRuntimeClassPathProvider)
@@ -32,15 +37,24 @@ class AnvilRuntimeEnvironmentConfigurator(testServices: TestServices) :
         configuration.addJvmClasspathRoot(file)
       }
     }
+    if (MetroDirectives.WITH_KI_ANVIL in module.directives) {
+      for (file in kiAnvilRuntimeClasspath) {
+        configuration.addJvmClasspathRoot(file)
+      }
+    }
   }
 }
 
 class AnvilRuntimeClassPathProvider(testServices: TestServices) :
   RuntimeClasspathProvider(testServices) {
   override fun runtimeClassPaths(module: TestModule): List<File> {
-    return when (MetroDirectives.WITH_ANVIL in module.directives) {
-      true -> anvilRuntimeClasspath
-      false -> emptyList()
+    return buildList {
+      if (MetroDirectives.WITH_ANVIL in module.directives) {
+        addAll(anvilRuntimeClasspath)
+      }
+      if (MetroDirectives.WITH_KI_ANVIL in module.directives) {
+        addAll(kiAnvilRuntimeClasspath)
+      }
     }
   }
 }